move event notification
