+++ spec/rest/webhooks/parse_spec.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
 require File.dirname(__FILE__) + '/../../spec_helper'
 
-describe SendGrid4r::REST::Webhooks::ParseApi do
+describe SendGrid4r::REST::Webhooks::Parse do
   describe 'integration test', :it do
     before do
       begin
@@ -18,7 +18,7 @@ describe SendGrid4r::REST::Webhooks::ParseApi do
         begin
           parse_settings = @client.get_parse_settings
           expect(parse_settings).to be_a(
-            SendGrid4r::REST::Webhooks::ParseApi::ParseSettings
+            SendGrid4r::REST::Webhooks::Parse::ParseSettings
           )
           expect(parse_settings.result).to be_a(Array)
         rescue RestClient::ExceptionWithResponse => e
@@ -62,16 +62,16 @@ describe SendGrid4r::REST::Webhooks::ParseApi do
       allow(client).to receive(:execute).and_return(parse_settings)
       actual = client.get_parse_settings
       expect(actual).to be_a(
-        SendGrid4r::REST::Webhooks::ParseApi::ParseSettings
+        SendGrid4r::REST::Webhooks::Parse::ParseSettings
       )
     end
 
     it 'creates parse_setting instance' do
-      actual = SendGrid4r::REST::Webhooks::ParseApi.create_parse_setting(
+      actual = SendGrid4r::REST::Webhooks::Parse.create_parse_setting(
         parse_setting
       )
       expect(actual).to be_a(
-        SendGrid4r::REST::Webhooks::ParseApi::ParseSetting
+        SendGrid4r::REST::Webhooks::Parse::ParseSetting
       )
       expect(actual.url).to eq('http://mydomain.com/parse')
       expect(actual.hostname).to eq('mail.mydomain.com')
@@ -79,11 +79,11 @@ describe SendGrid4r::REST::Webhooks::ParseApi do
     end
 
     it 'creates parse_settings instance' do
-      actual = SendGrid4r::REST::Webhooks::ParseApi.create_parse_settings(
+      actual = SendGrid4r::REST::Webhooks::Parse.create_parse_settings(
         parse_settings
       )
       expect(actual).to be_a(
-        SendGrid4r::REST::Webhooks::ParseApi::ParseSettings
+        SendGrid4r::REST::Webhooks::Parse::ParseSettings
       )
       expect(actual.result).to be_a(Array)
     end
